+++ src/main/java/com/Etech/Controller/ProductController.java
@@ -1,8 +1,7 @@
 package com.Etech.Controller;
 
+
 import com.Etech.Dto.ProductDto;
-import com.Etech.Exception.ResourceException;
-import com.Etech.Model.enums.ProductCategory;
 import com.Etech.Service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -16,33 +15,35 @@ import java.util.List;
 public class ProductController {
 
     @Autowired
-    private ProductService productService;
+    ProductService productService;
 
 
-    @GetMapping("search/{name}")
-    public ResponseEntity<List<ProductDto>> searchProducts(@PathVariable String name) {
-        try {
-            List<ProductDto> products = productService.searchProducts(name);
-            return new ResponseEntity<>(products, HttpStatus.OK);
-        } catch (ResourceException e) {
-            return new ResponseEntity<>(HttpStatus.valueOf(e.getStatus().value()));
-        } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
-        }
+    @GetMapping("products")
+    public ResponseEntity<List<ProductDto>> getAllProducts() {
+        return ResponseEntity.status( HttpStatus.OK).body(productService.getAllProducts());
     }
 
+    @GetMapping("searchByName/{name}")
+    public ResponseEntity<ProductDto> findProductByName(@PathVariable("name") String name) {
+        return ResponseEntity.status(HttpStatus.OK).body(productService.findProductByName (name));
+    }
     @GetMapping("searchByCategory/{category}")
-    public ResponseEntity<List<ProductDto>> searchProductsByCategory(@PathVariable String category) {
-        try {
-            ProductCategory cat = ProductCategory.valueOf(category.toUpperCase());
-            List<ProductDto> products = productService.findByCategory(cat);
-            return new ResponseEntity<>(products, HttpStatus.OK);
-        } catch (IllegalArgumentException e) {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        } catch (ResourceException e) {
-            return new ResponseEntity<>(HttpStatus.valueOf(e.getStatus().value()));
-        } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
-        }
+    public ResponseEntity<List<ProductDto>> findProductsByCategory(@PathVariable("category") String category) {
+        return ResponseEntity.status(HttpStatus.OK).body(productService.findAllByCatagory (category));
     }
+
+    @GetMapping("searchByKeyword/{keyword}")
+    public ResponseEntity<List<ProductDto>> findProductsByKeyword(@PathVariable("keyword") String keyword) {
+        return ResponseEntity.status(HttpStatus.OK).body(productService.findAllByKeyWord (keyword));
+    }
+
+//    @GetMapping("filterProducts")
+//    public ResponseEntity<List<ProductDto>> filterProducts(@RequestParam(name = "name", required = false) String name,
+//                                                           @RequestParam(name = "category", required = false) String category,
+//                                                           @RequestParam(name = "keyWord", required = false) String keyWord){
+//        return ResponseEntity.status(HttpStatus.OK).body(productService.filterProducts (name,category, keyWord));
+//    }
+
+
+
 }
