+++ src/main/java/com/Etech/Repository/ProductRepo.java
@@ -15,25 +15,11 @@ import java.util.Optional;
 public interface ProductRepo extends JpaRepository<Product, Long> {
 
     Optional<Product> findProductById(long id);
-
-
-//    @Query("SELECT p FROM Product p WHERE p.name= : name")
-      Optional<Product> getProductByName(@Param ("name") String name);
-//      @Query("SELECT p FROM Product p WHERE p.productCategory= : catagory")
-      List<Product> findProductsByProductCategory(@Param("catagory") ProductCategory category);
-
-//    @Query("SELECT p FROM Product p WHERE p.name LIKE %:keyWord%")
-
+    Optional<Product> getProductByName(@Param ("name") String name);
+    List<Product> findProductsByProductCategory(@Param("catagory") ProductCategory category);
     @Query("SELECT p FROM Product p WHERE LOWER(p.name) LIKE LOWER(CONCAT('%', :keyWord, '%'))")
     List<Product> findAllByKeyWord(@Param("keyWord") String keyWord);
-
-    // List<Product> findProductsWithStatus(@Param("") ProductStatus productStatus);
-
-
-
-
     List<Product> findByNameContainingIgnoreCase(String name);
-
     List<Product> findByProductCategory(ProductCategory category);
