+++ src/main/java/com/Etech/Service/ProductService.java
@@ -1,12 +1,19 @@
 package com.Etech.Service;
 
+
 import com.Etech.Dto.ProductDto;
-import com.Etech.Model.enums.ProductCategory;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Service
 public interface ProductService {
-    public List<ProductDto> searchProducts(String name);
 
-    public List<ProductDto> findByCategory(ProductCategory category);
+    ProductDto findProductByName(String name);
+    List<ProductDto> getAllProducts();
+    List<ProductDto> findAllByCatagory(String category);
+
+    List<ProductDto> findAllByKeyWord(String keyWord);
+
+//    List<ProductDto> filterProducts(String name, String category, String keyWord);
 }
